+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Chg #101: Rename `UserAuth` to `WebAuth`, mark `UserAuth` as deprecated (@olegbaturin)
 - New #101: Add `ApiAuth` authentication method (@olegbaturin)
+- Chg #109: Change PHP constraint in `composer.json` to `~8.1.0 || ~8.2.0 || ~8.3.0 || ~8.4.0` (@vjik)
+- Bug #109: Explicitly mark nullable parameters (@vjik)
+- Enh #109: Improve debug message "Unable to authenticate user…" in `LogginMiddleware` (@vjik) 
 
 ## 2.2.0 May 07, 2024